UBUNTU: SAUCE: platform: tegra_prod: Use kmemleak_not_leak for allocated memory
commit 7bedb618e348b6d0fe74c6210052ab434735323c from
https://nv-tegra.nvidia.com/r/3rdparty/canonical/linux-jammy.git
l4t/l4t-r36.2-Ubuntu-nvidia-tegra-1018.18

BugLink: https://bugs.launchpad.net/bugs/2028400

Some of prod setting is applied in the probe of the driver
and need to keep handle for future use. On this case, the
allocated memory does not get used for longer time and reported
as memory leak.

Add flag on allocated memory to say as this is not memleak.

Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Tested-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>