UBUNTU: SAUCE: PCI: tegra194: Allow system suspend when the Endpoint link is not up
commit 11d7f87c2a88db68a84e9c7f3477c6339e5d24ce from
https://nv-tegra.nvidia.com/r/3rdparty/canonical/linux-jammy.git
l4t/l4t-r36.2-Ubuntu-nvidia-tegra-1018.18

BugLink: https://bugs.launchpad.net/bugs/2001557

Only a Root port initiates the L2 sequence. PCIe link is kept in L2 state
during suspend. If Endpoint mode is enabled and the link is up, the
software cannot proceed with suspend. However, when the PCIe Endpoint
driver is probed, but the PCIe link is not up, Tegra can go into suspend
state. So, allow system to suspend in this case.

Signed-off-by: Vidya Sagar <vidyas@nvidia.com>
Signed-off-by: Kartik <kkartik@nvidia.com>
Acked-by: Brad Figg <bfigg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Brad Figg <bfigg@nvidia.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>