UBUNTU: SAUCE: dmaengine: tegra: Get slave channel based on DT
commit dcee68783ef433b364f0929cd88016381d8f6163 from
https://nv-tegra.nvidia.com/r/3rdparty/canonical/linux-jammy.git
l4t/l4t-r36.2-Ubuntu-nvidia-tegra-1018.18

BugLink: https://bugs.launchpad.net/bugs/2001557

Add option for DMA clients to provide channel number and
stream-id in device tree. dma-cells need to be increased to 3
for accommodating additional details.

This is required in a virtualized environment where all channels
and clients are not allowed use the same stream id.

The channel number read from device tree is modified to match with
the hardware channel number in the driver. This is done to avoid
ABI breakage. Otherwise, it would require updating all the device
trees that provide this information

Bug 200761027

Signed-off-by: Akhil R <akhilrajeev@nvidia.com>
Signed-off-by: Kartik <kkartik@nvidia.com>
Acked-by: Brad Figg <bfigg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Brad Figg <bfigg@nvidia.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>