UBUNTU: SAUCE: gpio: timestamping: program timestamp bit
commit 81cf500c0be205606dd5edc093087fa977e6c204 from
https://nv-tegra.nvidia.com/r/3rdparty/canonical/linux-jammy.git
l4t/l4t-r36.2-Ubuntu-nvidia-tegra-1018.18

BugLink: https://bugs.launchpad.net/bugs/2019240

There is a separate GTE (generic timestamping engine) driver in the
staging which are being used across auto and robotics teams. However
in addition to the GTE driver, AON GPIO controller still needs to enable
bit 7 to actually ungate timestamp functionality. With the lack of proper
callback to enable this bit from the GTE, we have to enable it by default
from the gpio driver instead. Ideally we should be enabling for the
requested gpio line only at run time but that is not possible with the
staging GTE driver.

Signed-off-by: Prathamesh Shete <pshete@nvidia.com>
Reviewed-by: Dipen Patel <dipenp@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Krishna Yarlagadda <kyarlagadda@nvidia.com>
Tested-by: Krishna Yarlagadda <kyarlagadda@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Brad Figg <bfigg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Brad Figg <bfigg@nvidia.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>