UBUNTU: SAUCE: Revert "UBUNTU: SAUCE: memory: tegra: Remove MC-SID override programming"
commit cf715e50b6923355ed5b7714339dd3ba71b66d6f from
https://nv-tegra.nvidia.com/r/3rdparty/canonical/linux-jammy.git
l4t/l4t-r36.2-Ubuntu-nvidia-tegra-1018.18

BugLink: https://bugs.launchpad.net/bugs/2028400

This reverts commit 6b01eca640e8a9fc15d5f54704e238101137afaf

Restore function tegra186_mc_probe_device() which will be called when
a device is attached with IOMMU. This function will match the client ID
and call SID override function for that client.
This is required as display SID is getting programmed to bypass in
bootloader and it need to overridden with correct SID in kernel.

Signed-off-by: Ashish Mhetre <amhetre@nvidia.com>
Reviewed-by: Pritesh Raithatha <praithatha@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>