UBUNTU: SAUCE: usb: typec: ucsi: Find connector fwnode without relying on order
commit 995b55beb826390da97bd70fc5c4309a6d25d035 from
https://nv-tegra.nvidia.com/r/3rdparty/canonical/linux-jammy.git
l4t/l4t-r36.2-Ubuntu-nvidia-tegra-1018.18

BugLink: https://bugs.launchpad.net/bugs/2028400

The previous code relied on the order of "connector" nodes under the
ucsi device node to determine the correct fwnode handle for a connector.
However, the order of nodes should not be relied upon when parsing device
trees, especially when using live DT overlays.

This change modifies ucsi_find_fwnode() to find the correct connector
fwnode by reading the "reg" property of each connector node and
comparing it to the connector number. If only a single connector node
exists without a "reg" property, that single node is returned.

This ensures the correct connector fwnode is found regardless of node
order in the DT.

Signed-off-by: Wayne Chang <waynec@nvidia.com>
Reviewed-by: Sing-Han Chen <singhanc@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>