NVIDIA: SAUCE: s25fs: Add post-get-map-id fixup for S25FS512S
commit d8a20a8fa1252b725c00f6b903de7dff8be2458d from
https://nv-tegra.nvidia.com/r/3rdparty/canonical/linux-jammy.git
l4t/l4t-r36.2-Ubuntu-nvidia-tegra-1018.18

BugLink: https://bugs.launchpad.net/bugs/2034622

Add a post-get-map-id fixup for S25FS512S flash devices. The fixup
reads the values of the CR3V and CR1V registers and determines the map ID
based on those values. The fixup also checks for invalid combinations of
CR3V and CR1V values and returns 0 in those cases.

This fixup is necessary to workaround an issue with the SFDP Address Map
for S25FS512S flash devices. The SFDP Address Map for these devices
incorrectly reports that the map ID is 0 when it should be 1. This
issue can cause problems when trying to erase sectors on the flash device.

The fixup in this commit will force the flash device to use the first
sector address map, which is the correct map for S25FS512S flash devices.

http://nvbugs/4218915

Signed-off-by: zuyih <zuyih@nvidia.com>
Signed-off-by: Vishwaroop A <va@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>